Make external module shutdown deterministic on Python 3

+++ shinken/basemodule.py
@@ -192,19 +192,15 @@ class BaseModule(object):
 
 
     def __kill(self):
-        """Sometime terminate() is not enough, we must "help"
-        external modules to die...
-        """
+        """Terminate an external module process and wait for it to be reaped."""
+        if self.process is None:
+            return
 
-        if os.name == 'nt':
-            self.process.terminate()
-        else:
-            # Ok, let him 1 second before really KILL IT
-            os.kill(self.process.pid, signal.SIGTERM)
-            time.sleep(1)
-            # You do not let me another choice guy...
-            if self.process.is_alive():
-                os.kill(self.process.pid, signal.SIGKILL)
+        self.process.terminate()
+        self.process.join(timeout=1)
+        if self.process.is_alive() and os.name != 'nt':
+            os.kill(self.process.pid, signal.SIGKILL)
+            self.process.join(timeout=1)
 
 
     def stop_process(self):
